Fix pert_labels type error in embedding layer
Convert pert_labels to long type before passing to nn.Embedding layer.
The embedding layer expects integer indices (Long/Int) but was receiving
FloatTensor, causing RuntimeError during forward pass.

+++ sedd/model.py
@@ -474,7 +474,7 @@ class SEDDPerturbationTransformer(SEDDTransformer):
         t_emb = self.time_embed(sigma)
 
         # Perturbation embedding
-        p_emb = self.pert_embed(pert_labels)
+        p_emb = self.pert_embed(pert_labels.long())
         p_emb = self.pert_proj(p_emb)
 
         # Combined conditioning: time + perturbation
